// types.ts
export enum Karat {
  K18 = 18,
  K21 = 21,
  K24 = 24,
}
// ...
export interface BullionItem {
  id: string;
  company: string;
  bullionType: string;
  weight: number;
  karat: Karat;
  count: number;
  totalWeight: number;
  shop?: string;
  updatedAt?: number;
}

export interface BullionLog {
  id: string;
  bullionId: string;
  company: string;
  bullionType: string;
  weight: number;
  operation: string;
  countChange: number;
  weightChange: number;
  isInOut: 'IN' | 'OUT';
  timestamp: number;
  shop?: string;
  updatedAt?: number;
}